// src/types/myAccount.tsx
interface IAccount {
  NomEntreprise: string;
  Pseudoname: string;
  Prenom: string;
  NomDeFamille: string;
  NumeroTelephone: string;
  Pays: string;
  Ville: string;
  AddresseComplete: string;
  Description: string;
  MinutePourSuprimer: string;
  MariageValeur: string;
  MariageQuantite: string;
}

export type { IAccount };